requirements.txt
hydra-core==1.1
einops==0.4.1
flow_vis==0.1
imageio==2.21.1
matplotlib==3.5.3
munch==2.5.0
numpy==1.23.5
omegaconf==2.1.0
opencv_python==4.6.0.66
opt_einsum==3.3.0
Pillow==9.5.0
pytorch_lightning==1.6.0
requests
scikit_image==0.19.2
scipy==1.9.1
setuptools==65.6.3
tabulate==0.8.10
tqdm==4.64.1
moviepy
